fix: apply the adversarial review's confirmed findings
A 36-agent review over this branch confirmed 22 findings before it shipped; the
ones that survive as code changes:

BLOCKING - the fingerprint was computed at hint-dequeue time, before
stabilization. The first hint for a cross-volume Bridge copy arrives at the
START of the copy, so the key was minted over a partial length and in-flight
mtime; the pipeline then patiently waited out the real copy and completed under
the stale key - after which the sweep computed the settled fingerprint, found no
receipt, and ran the entire import a second time: duplicate setlist entry,
duplicate cue, on essentially every large arrival. Three lenses found this
independently. The watcher now settles the file's identity first (size still,
no writer) and re-checks it after the run.

- Idle no longer lies during instrument setup: the wire reads Menu/NoSong there,
  byte-identical to idle, but the process holds evidence the wire lacks - a cue
  in pending-players. ObserveAsync now consults it, so a mid-party arrival can
  no longer click Scan Songs into the players' instrument selection.
- The cue command id derives from the import's idempotency key, so a
  crash-retry replays through the cue journal instead of re-driving the search
  box.
- A scene change during the scan settle is now Failed (retryable) rather than
  Ambiguous (held for eyes forever) - a party starting a song must not wedge an
  import permanently.
- A recovered-ambiguous setlist insert converges: if the entry is present it is
  AlreadyApplied; if absent it retries once under a recovery-suffixed id -
  instead of rejecting forever through a burnt command id.
- The reconciliation sweep is observed and survives non-IO exceptions; its death
  is named. The import loop's catch is broad, because an unnamed escape killed
  the consumer silently. ForgetFailure runs once per sweep, not once per queued
  filesystem event.
- A missing watch directory is waited for, once a minute, rather than disabling
  acquisition for the process lifetime while the log promises recovery.
- Containment no longer rejects everything when the watch directory carries a
  trailing separator (Path.GetFullPath preserves it; both prongs then failed).
- SngDocument's bounds check uses long arithmetic - a crafted length near
  int.MaxValue wrapped the int check and turned a named rejection into an
  exception that would have stopped the host at startup via the library scan.
- The library scan survives UnauthorizedAccessException (not an IOException) on
  both the ini and sng paths, for the same stops-the-host reason.
- The acquisition journal repairs a torn tail on open, parses unknown receipt
  outcomes as claimable rather than refusing to boot, rolls back an in-memory
  lease when its write never reached disk, and releases the lease on a failed
  finalize rather than wedging the key as InProgress until restart.
- FileArrivalPort clears its stability baseline on every terminal path rather
  than leaking one entry per never-stabilized file.
- LibraryIndexPort resolves by exact location rather than a title search.
- ClickAt reports SetCursorPos's actual result rather than an unconditional true.

Verified live after the fixes: Barkeep restarted against the real watch folder,
the settled fingerprint matched the recorded receipt, the import replayed as
Terminal, and the setlist still holds exactly one Everlong.

Accepted and documented rather than changed: the full-library rescan per arrival
(89 ms over 448 songs), the single serial import consumer (one operator, one
theater), and the pending-players -> Succeeded cue mapping (the import's promise
is play-next; the cue surface holds the true cue state).

Server tests 122 -> 126.

+++ src/Cantina.Barkeep/Library/SngDocument.cs
@@ -156,7 +156,11 @@ public sealed record SngDocument
         var length = BinaryPrimitives.ReadInt32LittleEndian(buffer.AsSpan(offset, 4));
         offset += 4;
 
-        if (length < 0 || offset + length > buffer.Length)
+        // Long arithmetic, deliberately: a crafted length near int.MaxValue wraps
+        // `offset + length` negative in int math, sails past the guard, and turns a
+        // named rejection into a thrown exception - which would have stopped the host
+        // at startup through the library scan. Found by review.
+        if (length < 0 || offset + (long)length > buffer.Length)
         {
             return false;
         }
